added attachments rendering to feed
