+++ less/com/message-summary.less
@@ -40,6 +40,16 @@
       padding-left: 2em;
     }
   }
+  .attachments {
+    padding: 0 0 1em;
+    a {
+      margin-right: 1em;
+      img {
+        max-width: 200px;
+        max-height: 200px;
+      }
+    }
+  }
   :not(.markdown) h4 {
     font-size: 15px;
     margin-top: 3px;
