+++ src/com/index.js
@@ -194,6 +194,7 @@ exports.message = require('./message')
 exports.messageThread = require('./message-thread')
 exports.messageSummary = require('./message-summary')
 exports.messageFeed = require('./message-feed')
+exports.messageAttachments = require('./message-attachments')
 exports.notifications = require('./notifications')
 exports.address = require('./address')
 exports.peers = require('./peers')
