+++ src/com/message-thread.js
@@ -1,6 +1,5 @@
 'use strict'
 var h = require('hyperscript')
-var mlib = require('ssb-msgs')
 var com = require('./index')
 var u = require('../lib/util')
 var markdown = require('../lib/markdown')
@@ -21,40 +20,13 @@ function getContent (app, msg) {
   } catch (e) { }
 }
 
-function getAttachments (app, msg) {
-  var imageTypes = {
-    png: 'image/png',
-    jpg: 'image/jpeg',
-    jpeg: 'image/jpeg',
-    gif: 'image/gif',
-    svg: 'image/svg+xml'  
-  }
-  function isImage (link) {
-    if (link.type && link.type.indexOf('image/') !== -1)
-      return true
-    if (link.name && imageTypes[link.name.split('.').slice(-1)[0].toLowerCase()])
-      return true
-  }
-
-  var els = []
-  mlib.indexLinks(msg.value.content, { ext: true }, function (link, rel) {
-    var label
-    if (isImage(link))
-      label = h('img', { src: '/ext/'+link.ext, title: link.name || link.ext })
-    else
-      label = [com.icon('file'), ' ', link.name, ' ', h('small', (('size' in link) ? u.bytesHuman(link.size) : ''), ' ', link.type||'')]
-    els.push(h('a', { href: '/ext/'+link.ext, target: '_blank' }, label))
-  })
-  return els
-}
-
 module.exports = function (app, thread, opts) {
 
   // markup
   
   var content = getContent(app, thread) || h('table', com.prettyRaw.table(app, thread.value.content))
   var viz = com.messageVisuals(app, thread)
-  var attachments = getAttachments(app, thread)
+  var attachments = com.messageAttachments(app, thread)
 
   opts.onRender && opts.onRender(thread)
